requirements.txt
absl-py==1.0.0
astunparse==1.6.3
backcall==0.2.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
entrypoints==0.3
flatbuffers==2.0
fonttools==4.28.2
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.1
h5py==3.5.0
idna==3.3
ipykernel==6.5.1
ipython==7.29.0
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.0.6
jupyter-core==4.9.1
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
kneed==0.7.0
libclang==12.0.0
Markdown==3.3.4
matplotlib==3.5.0
matplotlib-inline==0.1.3
nest-asyncio==1.5.1
numpy==1.21.4
oauthlib==3.1.1
opencv-python==4.5.4.60
opt-einsum==3.3.0
packaging==21.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.22
protobuf==3.19.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.10.0
pyparsing==3.0.6
python-dateutil==2.8.2
pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==1.0.1
scipy==1.7.2
setuptools-scm==6.3.2
six==1.16.0
sklearn==0.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
threadpoolctl==3.0.0
tomli==1.2.2
tornado==6.1
traitlets==5.1.1
typing-extensions==3.10.0.2
urllib3==1.26.7
wcwidth==0.2.5
Werkzeug==2.0.2
wrapt==1.13.3
